(Fix) 메인 버튼 클릭시 페이지 전환 오류 문제 해결
