+++ src/pages/login.html
@@ -27,7 +27,7 @@
                 </div>
                 <div class="helper-text" id="helper-text"></div>
             </article>
-            <button class="main-btn" id="login-btn" >
+            <button class="main-btn" id="login-btn" type="button">
                 로그인</button>
             <a onclick="location.href='./signin.html'" class="basic-a-link">
                 회원가입</a>
@@ -35,6 +35,6 @@
         </form>
     </main>
     <script src="../components/header.js"></script>
-    <script src = "login.js"></script>
+    <script type="module" src = "login.js"></script>
 </body>
 </html>
